requirements.txt
flask>=2.0.0
pywebview>=4.0.0
pyinstaller>=5.0.0